main/requirements.txt
numpy
tqdm
face_recognition
Pillow